chore: clean usage of `skipLibCheck` and `skipDefaultLibCheck` TS option
`skipDefaultLibCheck` is deprecated so there is no need to add it
anymore.

I added `skipLibCheck` in the angular example because it looks like
there are some type incompatibility between some types coming from
node_modules we have in the monorepo. (Not really a surprise with the
number of deps we have)

Signed-off-by: Gaëtan Maisse <gaetanmaisse@gmail.com>